delete api router fixed

+++ src/routes/deleteCollection.ts
@@ -3,9 +3,15 @@ import { qdrant } from "../lib/qdrant";
 
 export const router = Router();
 
-router.delete('/',async(res:Response, req:Request)=>{
+router.delete('/',async( req:Request ,res:Response)=>{
     try{
         const {collectionName}=req.body;
+        console.log("collection name-", collectionName);
+        if(!collectionName){
+            res.status(400).json({message:"please enter collection name"});
+            return;
+        }
+
         const response =await qdrant.deleteCollection(`${collectionName}`)
         console.log("delete collection route response- ", response);
